+++ src/test/java/li/strolch/model/timedstate/TimeStateTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import li.strolch.model.timevalue.ITimeValue;
 import li.strolch.model.timevalue.IValueChange;
-import li.strolch.model.timevalue.impl.DoubleValue;
+import li.strolch.model.timevalue.impl.FloatValue;
 import li.strolch.model.timevalue.impl.ValueChange;
 
 import org.junit.Before;
@@ -28,10 +28,10 @@ import org.junit.Test;
 
 public class TimeStateTest {
 
-	private ITimedState<DoubleValue> state = new TimedState<DoubleValue>();
+	private ITimedState<FloatValue> state = new TimedState<FloatValue>();
 
-	final DoubleValue expectedValue1 = new DoubleValue(Double.valueOf(100D));
-	final DoubleValue expectedValue2 = new DoubleValue(Double.valueOf(200D));
+	final FloatValue expectedValue1 = new FloatValue(Double.valueOf(100D));
+	final FloatValue expectedValue2 = new FloatValue(Double.valueOf(200D));
 
 	final Long t0 = Long.valueOf(0);
 	final Long t10 = Long.valueOf(10);
@@ -42,20 +42,20 @@ public class TimeStateTest {
 	@Before
 	public void before() {
 
-		final IValueChange<DoubleValue> change1 = new ValueChange<DoubleValue>(this.t10, this.expectedValue1);
+		final IValueChange<FloatValue> change1 = new ValueChange<FloatValue>(this.t10, this.expectedValue1);
 		this.state.applyChange(change1);
 
-		final ITimeValue<DoubleValue> stateAt9 = this.state.getStateAt(9L);
+		final ITimeValue<FloatValue> stateAt9 = this.state.getStateAt(9L);
 		assertNull(stateAt9);
 
-		final ITimeValue<DoubleValue> stateAt11 = this.state.getStateAt(11L);
+		final ITimeValue<FloatValue> stateAt11 = this.state.getStateAt(11L);
 		assertNotNull(stateAt11);
 		assertEquals(true, stateAt11.getValue().matches(this.expectedValue1));
 
-		final IValueChange<DoubleValue> change2 = new ValueChange<DoubleValue>(this.t30, this.expectedValue1);
+		final IValueChange<FloatValue> change2 = new ValueChange<FloatValue>(this.t30, this.expectedValue1);
 		this.state.applyChange(change2);
 
-		final ITimeValue<DoubleValue> stateAt31 = this.state.getStateAt(31L);
+		final ITimeValue<FloatValue> stateAt31 = this.state.getStateAt(31L);
 		assertNotNull(stateAt31);
 		assertEquals(true, stateAt31.getValue().matches(this.expectedValue2));
 	}
@@ -63,7 +63,7 @@ public class TimeStateTest {
 	@Test
 	public void testGetNextMatch() {
 
-		ITimeValue<DoubleValue> nextMatch = this.state.getNextMatch(this.t0, this.expectedValue1);
+		ITimeValue<FloatValue> nextMatch = this.state.getNextMatch(this.t0, this.expectedValue1);
 		assertNotNull(nextMatch);
 		assertEquals(this.t10, nextMatch.getTime());
 
@@ -89,7 +89,7 @@ public class TimeStateTest {
 	@Test
 	public void testGetPreviousMatch() {
 
-		ITimeValue<DoubleValue> previousMatch = this.state.getPreviousMatch(this.t100, this.expectedValue2);
+		ITimeValue<FloatValue> previousMatch = this.state.getPreviousMatch(this.t100, this.expectedValue2);
 		assertNotNull(previousMatch);
 		assertEquals(this.t30, previousMatch.getTime());
