+++ src/test/java/li/strolch/model/timevalue/FloatTimeVariableTest.java
@@ -22,7 +22,7 @@ import static org.junit.Assert.assertTrue;
 import java.util.Collection;
 import java.util.SortedSet;
 
-import li.strolch.model.timevalue.impl.DoubleValue;
+import li.strolch.model.timevalue.impl.FloatValue;
 import li.strolch.model.timevalue.impl.TimeVariable;
 import li.strolch.model.timevalue.impl.ValueChange;
 
@@ -35,22 +35,22 @@ public class FloatTimeVariableTest {
 	private static final Long STEP = 10L;
 	private static final Long PICK = 50L;
 
-	private TimeVariable<DoubleValue> timeVariable;
+	private TimeVariable<FloatValue> timeVariable;
 
 	/**
 	 * set the values ascending with a difference of STEP
 	 */
 	@Before
 	public void init() {
-		this.timeVariable = new TimeVariable<DoubleValue>();
+		this.timeVariable = new TimeVariable<FloatValue>();
 		for (long i = 0; i < MAX; i += STEP) {
-			this.timeVariable.setValueAt(Long.valueOf(i), new DoubleValue(i));
+			this.timeVariable.setValueAt(Long.valueOf(i), new FloatValue(i));
 		}
 	}
 
 	@Test
 	public void testGetValueAt() {
-		ITimeValue<DoubleValue> valueAt = this.timeVariable.getValueAt(PICK);
+		ITimeValue<FloatValue> valueAt = this.timeVariable.getValueAt(PICK);
 		assertEquals(PICK.doubleValue(), valueAt.getValue().getValue(), 0.0001);
 	}
 
@@ -59,12 +59,12 @@ public class FloatTimeVariableTest {
 	 */
 	@Test
 	public void testGetFutureValues() {
-		Collection<ITimeValue<DoubleValue>> futureValues = this.timeVariable.getFutureValues(PICK);
+		Collection<ITimeValue<FloatValue>> futureValues = this.timeVariable.getFutureValues(PICK);
 		Long expectedTime = PICK;
 		Double expectedValue = PICK.doubleValue();
-		for (ITimeValue<DoubleValue> value : futureValues) {
+		for (ITimeValue<FloatValue> value : futureValues) {
 			assertEquals(expectedTime, value.getTime());
-			assertTrue(value.getValue().matches(new DoubleValue(expectedValue)));
+			assertTrue(value.getValue().matches(new FloatValue(expectedValue)));
 			expectedTime += STEP;
 			expectedValue += STEP.doubleValue();
 		}
@@ -76,12 +76,12 @@ public class FloatTimeVariableTest {
 	 */
 	@Test
 	public void testGetPastValues() {
-		Collection<ITimeValue<DoubleValue>> pastValues = this.timeVariable.getPastValues(MAX);
+		Collection<ITimeValue<FloatValue>> pastValues = this.timeVariable.getPastValues(MAX);
 		Long expectedTime = 0L;
 		Double expectedValue = expectedTime.doubleValue();
-		for (ITimeValue<DoubleValue> value : pastValues) {
+		for (ITimeValue<FloatValue> value : pastValues) {
 			assertEquals(expectedTime, value.getTime());
-			assertTrue(value.getValue().matches(new DoubleValue(expectedValue)));
+			assertTrue(value.getValue().matches(new FloatValue(expectedValue)));
 			expectedTime += STEP;
 			expectedValue += STEP.doubleValue();
 		}
@@ -93,17 +93,17 @@ public class FloatTimeVariableTest {
 	@Test
 	public void testApplyChange() {
 
-		DoubleValue doubleValue = new DoubleValue(STEP.doubleValue());
+		FloatValue doubleValue = new FloatValue(STEP.doubleValue());
 
-		IValueChange<DoubleValue> change = new ValueChange<DoubleValue>(PICK, doubleValue);
+		IValueChange<FloatValue> change = new ValueChange<FloatValue>(PICK, doubleValue);
 		this.timeVariable.applyChange(change);
 
-		Collection<ITimeValue<DoubleValue>> futureValues = this.timeVariable.getFutureValues(PICK);
+		Collection<ITimeValue<FloatValue>> futureValues = this.timeVariable.getFutureValues(PICK);
 		Long expectedTime = PICK;
 
-		IValue<Double> expectedValue = new DoubleValue(PICK.doubleValue() + change.getValue().getValue());
+		IValue<Double> expectedValue = new FloatValue(PICK.doubleValue() + change.getValue().getValue());
 
-		for (ITimeValue<DoubleValue> value : futureValues) {
+		for (ITimeValue<FloatValue> value : futureValues) {
 			assertEquals(expectedTime, value.getTime());
 			assertTrue(expectedValue.matches(value.getValue()));
 			expectedTime += STEP;
@@ -117,17 +117,17 @@ public class FloatTimeVariableTest {
 	@Test
 	public void testApply2Change() {
 
-		this.timeVariable = new TimeVariable<DoubleValue>();
+		this.timeVariable = new TimeVariable<FloatValue>();
 
-		DoubleValue doubleValue = new DoubleValue(STEP.doubleValue());
+		FloatValue doubleValue = new FloatValue(STEP.doubleValue());
 
-		IValueChange<DoubleValue> change = new ValueChange<DoubleValue>(PICK, doubleValue);
+		IValueChange<FloatValue> change = new ValueChange<FloatValue>(PICK, doubleValue);
 		this.timeVariable.applyChange(change);
 
-		ITimeValue<DoubleValue> actual = this.timeVariable.getValueAt(PICK); 
+		ITimeValue<FloatValue> actual = this.timeVariable.getValueAt(PICK); 
 		assertNotNull(actual); 	
 		
-		IValue<Double> expectedValue = new DoubleValue(STEP.doubleValue());
+		IValue<Double> expectedValue = new FloatValue(STEP.doubleValue());
 		assertEquals(true, actual.getValue().matches(expectedValue));
 	}
 
@@ -138,14 +138,14 @@ public class FloatTimeVariableTest {
 	@Test
 	public void testCompact() {
 
-		this.timeVariable = new TimeVariable<DoubleValue>();
+		this.timeVariable = new TimeVariable<FloatValue>();
 		for (Long i = 0L; i < MAX; i += STEP) {
-			this.timeVariable.setValueAt(i, new DoubleValue(STEP.doubleValue()));
+			this.timeVariable.setValueAt(i, new FloatValue(STEP.doubleValue()));
 		}
 		// call
 		this.timeVariable.compact();
 		// check
-		SortedSet<ITimeValue<DoubleValue>> futureValues = this.timeVariable.getFutureValues(0L);
+		SortedSet<ITimeValue<FloatValue>> futureValues = this.timeVariable.getFutureValues(0L);
 		assertEquals(1, futureValues.size());
 	}
